tut12

+++ app/Http/Controllers/PostsController.php
@@ -3,11 +3,21 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 use DB;
+use Redirect;
+use Session;
 
 class PostsController extends Controller
 {
+
+    public function __construct()
+    {
+       //session is not working inside construct
+    }
+
+    
     /**
      * Display a listing of the resource.
      *
@@ -35,6 +45,10 @@ class PostsController extends Controller
      */
     public function create()
     {
+        if(empty(Session::get('id')) ){
+            return redirect('/users'); 
+        }
+
         return view('posts.create');
     }
 
@@ -48,14 +62,33 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        //Handle File Upload
+        if($request->hasFile('cover_image')){
+            //Get filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //Get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            //Get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            //Filename to store
+            $fileNameToStore = $filename.'_'.date('YmdHis').'.'.$extension;
+            //Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+
+        }else{
+            $fileNameToStore = 'noimage.jpg';
+        }
 
         //Create Post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->user_id = session('id');
+        $post->cover_image = $fileNameToStore;
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Created');
@@ -82,7 +115,17 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
+        if(empty(Session::get('id')) ){
+            return redirect('/users'); 
+        }
+
         $post =  Post::find($id);
+
+        //check for correct user
+        if( Session::get('id') !== $post->user_id ){
+            return redirect('/posts')->with('error', 'Unauthorized');
+        }
+
         return view('posts.edit')->with('post',$post);
     }
 
@@ -115,13 +158,32 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        //Handle File Upload
+        if($request->hasFile('cover_image')){
+            //Get filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //Get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            //Get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            //Filename to store
+            $fileNameToStore = $filename.'_'.date('YmdHis').'.'.$extension;
+            //Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+
+        }
 
         //Update Post
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if($request->hasFile('cover_image')){
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Updated');
@@ -129,9 +191,23 @@ class PostsController extends Controller
 
     public function delete_custom($id)
     {
-        $post = Post::find($id);
-        $post->delete();
+        if(empty(Session::get('id')) ){
+            return redirect('/users'); 
+        }
 
+        $post = Post::find($id);
+
+        //check for correct user
+        if( Session::get('id') !== $post->user_id ){
+            return redirect('/posts')->with('error', 'Unauthorized');
+        }
+
+        if($post->cover_image != 'noimage.jpg'){
+            //Delete Image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
+        
+        $post->delete();
         return redirect('/posts')->with('success', 'Post Removed');
     }
